# Minor: Log TPCH benchmark results